+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## Unreleased
 
-### Changed
+### Added
 
+  - Voucher report - create new voucher-report resource
+
+### Change
+
+  - Voucher - add supplier to voucher
   - Create user callback should not fail silently
 
 ## [4.2.2] - 2019-08-07